pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ollamadiffuser"
version = "1.0.0"
authors = [
    {name = "OllamaDiffuser Team", email = "ollamadiffuser@gmail.com"}
]
description = "🎨 Ollama-like image generation model management tool with LoRA support"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10"
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Multimedia :: Graphics",
]
keywords = ["diffusion", "image-generation", "ai", "machine-learning", "lora", "ollama"]
dependencies = [
    "torch>=2.1.0",
    "diffusers>=0.26.0",
    "transformers>=4.35.0",
    "accelerate>=0.25.0",
    "fastapi>=0.104.0",
    "uvicorn>=0.23.0",
    "huggingface-hub>=0.16.0",
    "Pillow>=9.0.0",
    "click>=8.0.0",
    "rich>=13.0.0",
    "pydantic>=2.0.0",
    "protobuf>=3.20.0",
    "sentencepiece>=0.1.99",
    "safetensors>=0.3.0",
    "python-multipart>=0.0.0",
    "psutil>=5.9.0",
    "jinja2>=3.0.0",
    "peft>=0.10.0",
    "controlnet-aux>=0.0.7",
    "opencv-python>=4.8.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "flake8>=6.0.0",
]

[project.scripts]
ollamadiffuser = "ollamadiffuser.__main__:main"

[project.urls]
Homepage = "https://github.com/ollamadiffuser/ollamadiffuser"
Repository = "https://github.com/ollamadiffuser/ollamadiffuser"
Issues = "https://github.com/ollamadiffuser/ollamadiffuser/issues"

[tool.setuptools]
include-package-data = true
zip-safe = false

[tool.setuptools.packages.find]
where = ["."]
include = ["ollamadiffuser*"]

# Optional: Black configuration
[tool.black]
line-length = 88
target-version = ['py310']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

# Optional: isort configuration
[tool.isort]
profile = "black"
multi_line_output = 3
line_length = 88
known_first_party = ["ollamadiffuser"]

# Optional: pytest configuration
[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --tb=short" 